Seperate out status generation.
